Fixing the `mix holidefs.download` task
It was actually a problem in the compilation process. As we were raising
an exception when there was no file for a given locale, which is the
case when we just added a new locale to the list and we're trying to use
`mix holidefs.download` to download that file.